Time: 1 ms (97.44%), Space: 39.6 MB (40.02%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -10,7 +10,7 @@ class Solution {
          result = mid;
          return result;
      }
-     else if((long) mid * mid > (long) x)
+     else if((long) mid * mid > x)
                 end = mid-1;
          
             else
